feat: Refactor artist page layout and restore metadata generation for Spotify artist details
