+++ src/app/(protected)/spotify/artists/[id]/loading.tsx
